PD-13507 revert useEscapeKey
