+++ packages/SidePanel/src/hooks/index.js
@@ -1,2 +1,3 @@
+export { default as useEscapeKey } from "./useEscapeKey";
 export { default as useOffsetScroll } from "./useOffsetScroll";
 export { default as useFooterOffset } from "./useFooterOffset";
